sqlite: code simplification / clean up
