+++ lib/Queries.js
@@ -2,10 +2,6 @@
 const MAX_RESULTS = 100;
 // const WILDCARD_LIMIT = 100000;
 
-function SqlDatabase( db ){
-  this.db = db;
-}
-
 var TIMER = false;
 var DEBUG = false;
 var AUTOCOMPLETE = false;
@@ -28,7 +24,7 @@ function singleQuoteEach( values ){
 }
 
 // cb( bool ) whether a 'subject' value exists in the db
-SqlDatabase.prototype.hasSubject = function( subject, cb ){
+module.exports.hasSubject = function( subject, cb ){
 
   var sql = [
     'SELECT id',
@@ -51,7 +47,7 @@ SqlDatabase.prototype.hasSubject = function( subject, cb ){
   });
 };
 
-SqlDatabase.prototype.hasSubjectAutocomplete = function( subject, cb ){
+module.exports.hasSubjectAutocomplete = function( subject, cb ){
 
   var sql = [
     'SELECT id',
@@ -77,7 +73,7 @@ SqlDatabase.prototype.hasSubjectAutocomplete = function( subject, cb ){
   });
 };
 
-SqlDatabase.prototype.matchSubjectDistinctSubjectIds = function( subject, cb ){
+module.exports.matchSubjectDistinctSubjectIds = function( subject, cb ){
   var sql = [
     'SELECT DISTINCT( t1.id ) AS subjectId',
     'FROM tokens AS t1',
@@ -102,7 +98,7 @@ SqlDatabase.prototype.matchSubjectDistinctSubjectIds = function( subject, cb ){
   });
 };
 
-SqlDatabase.prototype.matchSubjectAutocompleteDistinctSubjectIds = function( subject, cb ){
+module.exports.matchSubjectAutocompleteDistinctSubjectIds = function( subject, cb ){
 
   var sql = [
     'SELECT DISTINCT( t1.id ) AS subjectId',
@@ -131,7 +127,7 @@ SqlDatabase.prototype.matchSubjectAutocompleteDistinctSubjectIds = function( sub
   });
 };
 
-SqlDatabase.prototype.matchSubjectObject = function( subject, object, cb ){
+module.exports.matchSubjectObject = function( subject, object, cb ){
   var sql = [
     // 'SELECT t1.id AS subjectId, t1.token AS subject, t2.id as objectId, t2.token AS object',
     'SELECT t1.id AS subjectId, t2.id as objectId',
@@ -163,7 +159,7 @@ SqlDatabase.prototype.matchSubjectObject = function( subject, object, cb ){
   });
 };
 
-SqlDatabase.prototype.matchSubjectObjectAutocomplete = function( subject, object, cb ){
+module.exports.matchSubjectObjectAutocomplete = function( subject, object, cb ){
   var sql = [
     // 'SELECT t1.id AS subjectId, t1.token AS subject, t2.id as objectId, t2.token AS object',
     'SELECT t1.id AS subjectId, t2.id as objectId',
@@ -194,5 +190,3 @@ SqlDatabase.prototype.matchSubjectObjectAutocomplete = function( subject, object
     cb( null, rows );
   });
 };
-
-module.exports = SqlDatabase;
